fix(cli): base CLI crashed on default install without [ops] extra (fastapi)
`pip install attune-ai` (no `[ops]` extra) installs no fastapi, yet the
base `attune` CLI imported it transitively and crashed on startup with
`ModuleNotFoundError: No module named 'fastapi'` — `attune --help` and
every command died. Chain: cli_minimal -> cli_commands.curator ->
curator.sources.specs imported `SpecRecord` / `_list_specs_in_root`
from the FastAPI web-route module `attune.ops.routes.specs`.

CI never caught it because CI always installs dev/ops extras (fastapi
present), so the unit suite imported fine. Caught by a clean-venv
`attune --help` smoke of the shipped wheel during 8.6.0 release QA.
Pre-existing since #485 — published 8.5.0 has the identical crash.

Fix: extract the pure, framework-free spec-listing data layer
(SpecPhase, SpecRecord, _list_specs_in_root, and helpers) into a new
`attune.ops.specs_data` module. `routes/specs.py` re-exports the names
for backward compatibility (dashboard route + tests unchanged); the
curator imports from `specs_data`, so the base CLI no longer touches
fastapi.

Regression guard: tests/unit/test_base_cli_imports_without_ops_extra.py
imports cli_minimal + the curator spec source in a subprocess with
fastapi blocked — the exact no-[ops] environment CI's extras-installed
suite can't reproduce.

Verified: shipped 8.6.0 wheel installed in a clean fastapi-free venv —
`attune --help` exits 0, `attune version` -> "attune-ai 8.6.0"; route
re-exports still import with fastapi present; 163 affected tests pass.
